agent/requirements.txt
langchain>=0.3.25
langchain-mcp-adapters>=0.0.11
langchain-google-genai>=0.1.0
langgraph>=0.4.3
mcp-server-git>=2025.1.14
mcp[cli]>=1.8.0
pandas>=2.2.3
pydantic>=2.11.4
python-dotenv>=1.1.0
